fixed type problem

+++ trafficcounter-backend/src/main/java/ch/bfh/trafficcounter/model/dto/HistoricDataDto.java
@@ -12,6 +12,6 @@ public class HistoricDataDto {
 
     private int avgVehicleAmount;
 
-    private int avgVehicleSpeed;
+    private double avgVehicleSpeed;
 
 }
